+++ R/h_freq_funs.R
@@ -178,6 +178,20 @@ h_colexpr_substr <- function(var, col_expr) {
   return(col_expr_substr)
 }
 
+#' Identify if column is a combo column
+#'
+#'
+#'
+#' @param trtvar treatment Variable to extract from the expression.
+#' @param .spl_context from rtables machinery
+#' @return TRUE if in a combined column facet.
+#' @noRd
+#' @keywords internal
+h_in_combo <- function(trtvar, .spl_context){
+  colexpr_substr <- h_colexpr_substr(trtvar, .spl_context$cur_col_expr[[1]])
+  in_combo <- grepl(pattern = "c(", x = colexpr_substr, fixed = TRUE)
+}
+
 #' Get Denominator Parent Data Frame
 #'
 #' Retrieves the parent data frame based on denominator.