+++ R/junco_utils_default_stats_formats_labels.R
@@ -196,6 +196,7 @@ junco_default_stats <- list(
   summarize_ancova_j = c(
     "n",
     "mean_sd",
+    "mean_ci_3d",
     "median",
     "range",
     "quantiles",